refactor: rename indexes
