+++ src/games/ItGame/pages/pages.ts
