+++ src/games/ItGame/utils/constants.js
@@ -1,4 +1,1 @@
 export const baseUrl = 'https://myowngame.onrender.com/mog/api';
-export const apiHeaders = {
-  'Content-Type': ['application/json', 'text/plain', '*/*'],
-};
